# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EZKey API",
    "version": "0.4.0",
    "description": "Programmatic API for EZKey — request TLS certificates, SSH keys, and SSH certificates. Authenticate with an API token (`Authorization: Bearer ezk_…`) minted on the API Tokens page, or with a login JWT. API tokens are gated by fine-grained scopes; each endpoint below notes the scope it requires."
  },
  "servers": [
    { "url": "/api/v1", "description": "EZKey API base path" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ezk_ API token or login JWT",
        "description": "Send `Authorization: Bearer <token>`. Use an `ezk_…` API token for scripts/CI, or a JWT from POST /auth/login."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "token_type": { "type": "string", "example": "Bearer" },
          "expires_in_seconds": { "type": "integer" },
          "username": { "type": "string" },
          "role": { "type": "string" }
        }
      },
...
    }
  },
  "security": [ { "bearerAuth": [] } ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Obtain a login JWT",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["username", "password"], "properties": { "username": { "type": "string" }, "password": { "type": "string" } } } } }
        },
        "responses": {
          "200": { "description": "Token issued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenResponse" } } } },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}